# packages/api-contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "APMTRACK API",
    "version": "0.1.0",
    "description": "Contrat OpenAPI backend APMTRACK: sante, auth, RBAC, communes, utilisateurs, agents, zones, referentiel, PV, paiements, signalements, patrouilles, dashboard, audit logs et exports."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Developpement local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/exports/pvs": {
      "get": {
        "summary": "Exporte pvs en CSV",
        "responses": {
          "200": {
            "description": "Export CSV",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/exports/payments": {
      "get": {
        "summary": "Exporte payments en CSV",
        "responses": {
          "200": {
            "description": "Export CSV",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/exports/signalements": {
      "get": {
        "summary": "Exporte signalements en CSV",
        "responses": {
          "200": {
            "description": "Export CSV",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/exports/agents": {
      "get": {
        "summary": "Exporte agents en CSV",
        "responses": {
          "200": {
            "description": "Export CSV",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Authentification invalide",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Role ou commune non autorise",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "details"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            }
          }
        }
      },
...
    }
  }
}